LayoutTests/security/text-decode-long-strings-expected.txt

CONSOLE MESSAGE: error RangeError: Bad value
CONSOLE MESSAGE: error RangeError: Bad value
